[Notification] feat: 템플릿 컨버팅 로직 분리 #34
